Copy lightster/hodor to hold-the-door/ravens
Copy was made as of
lightster/hodor#fb1e15a43cb245d7996d9471a0aba43049d2eb53

Ravens is going to be the message queue part of Hodor.
I am splitting Hodor\MessageQueue from the rest of the
project so that it can be changed in smaller chunks
without constantly having to update the rest of Hodor.

Everything non-MessageQueue will be removed in a separate
PR so history can more easily show what changed between
Hodor and Ravens.